+++ product_code/exam_service/impl/src/main/scala/de/upb/cs/uc4/exam/impl/ExamServiceImpl.scala
@@ -35,7 +35,7 @@ class ExamServiceImpl(
     certificateService: CertificateService,
     operationService: OperationService,
     override val environment: Environment
-)(implicit ec: ExecutionContext, config: Config, materializer: Materializer)
+)(implicit ec: ExecutionContext, override val config: Config, materializer: Materializer)
   extends ExamService {
 
   /** Looks up the entity for the given ID */
@@ -110,7 +110,7 @@ class ExamServiceImpl(
                           case throwable: Throwable =>
                             log.error("Exception in addToWatchlist getProposalAddExam", throwable)
                         }
-                      (ResponseHeader(200, MessageProtocol.empty, List()), UnsignedProposal(proposalWrapper.proposal))
+                      (ResponseHeader(200, MessageProtocol.empty, List()), createTimedUnsignedProposal(proposalWrapper.proposal))
                   }.recover(handleException("getProposalAddExam failed"))
                 }
             }.recover {