+++ product_code/hyperledger_component/api/src/main/scala/de/upb/cs/uc4/hyperledger/api/model/SignedProposal.scala
@@ -1,12 +1,14 @@
 package de.upb.cs.uc4.hyperledger.api.model
 
+import de.upb.cs.uc4.hyperledger.api.JWTClaimParser
 import play.api.libs.json.{ Format, Json }
 
 import java.util.Base64
 
-case class SignedProposal(unsignedProposal: String, signature: String) {
+case class SignedProposal(unsignedProposalJwt: String, signature: String) {
 
-  def unsignedProposalAsByteArray: Array[Byte] = Base64.getDecoder.decode(unsignedProposal)
+  lazy val unsignedProposalAsByteArray: Array[Byte] =
+    Base64.getDecoder.decode(JWTClaimParser.readClaim("unsignedBytes", unsignedProposalJwt))
 
   def signatureAsByteArray: Array[Byte] = Base64.getDecoder.decode(signature)
 }