+++ product_code/hyperledger_component/api/src/main/scala/de/upb/cs/uc4/hyperledger/api/model/SignedTransaction.scala
@@ -1,12 +1,14 @@
 package de.upb.cs.uc4.hyperledger.api.model
 
+import de.upb.cs.uc4.hyperledger.api.JWTClaimParser
 import play.api.libs.json.{ Format, Json }
 
 import java.util.Base64
 
-case class SignedTransaction(unsignedTransaction: String, signature: String) {
+case class SignedTransaction(unsignedTransactionJwt: String, signature: String) {
 
-  def unsignedTransactionAsByteArray: Array[Byte] = Base64.getDecoder.decode(unsignedTransaction)
+  lazy val unsignedTransactionAsByteArray: Array[Byte] =
+    Base64.getDecoder.decode(JWTClaimParser.readClaim("unsignedBytes", unsignedTransactionJwt))
 
   def signatureAsByteArray: Array[Byte] = Base64.getDecoder.decode(signature)
 }