+++ product_code/operation_service/impl/src/main/scala/de/upb/cs/uc4/operation/impl/actor/RawUnsignedProposal.scala
@@ -1,0 +1,3 @@
+package de.upb.cs.uc4.operation.impl.actor
+
+case class RawUnsignedProposal(unsignedProposalBytes: Array[Byte])