+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/exceptions/UC4Exception.scala
@@ -29,6 +29,7 @@ object UC4Exception {
   val AlreadyDeleted = new UC4NonCriticalException(400, GenericError(ErrorType.AlreadyDeleted))
   val MalformedRefreshToken = new UC4NonCriticalException(400, GenericError(ErrorType.MalformedRefreshToken))
   val MalformedLoginToken = new UC4NonCriticalException(400, GenericError(ErrorType.MalformedLoginToken))
+  val MalformedFrontendSigningToken = new UC4NonCriticalException(400, GenericError(ErrorType.MalformedFrontendSigningToken))
   val MultipleAuthorizationError = new UC4NonCriticalException(400, GenericError(ErrorType.MultipleAuthorization))
   def QueryParameterError(invalidParams: SimpleError*) =
     new UC4NonCriticalException(400, DetailedError(ErrorType.QueryParameter, invalidParams))
@@ -55,6 +56,8 @@ object UC4Exception {
   val PathParameterMismatch = new UC4NonCriticalException(422, GenericError(ErrorType.PathParameterMismatch))
   val RefreshTokenSignatureError = new UC4NonCriticalException(422, GenericError(ErrorType.RefreshTokenSignatureInvalid))
   val LoginTokenSignatureError = new UC4NonCriticalException(422, GenericError(ErrorType.LoginTokenSignatureInvalid))
+  val FrontendSigningTokenSignatureError = new UC4NonCriticalException(422, GenericError(ErrorType.FrontendSigningTokenSignatureInvalid))
+  val FrontendSigningTokenExpired = new UC4NonCriticalException(422, GenericError(ErrorType.FrontendSigningTokenExpired))
   //428
   val PreconditionRequired = new UC4NonCriticalException(428, GenericError(ErrorType.PreconditionRequired))
   //500