+++ product_code/admission_service/CHANGELOG.md
@@ -1,3 +1,9 @@
+# [WIP - v0.19.2](https://github.com/upb-uc4/University-Credits-4.0/compare/admission-v0.19.1...admission-v0.19.2) (2021-XX-XX)
+## Feature
+- Added token with expiration date to every getProposal
+## Refactor
+## Bugfix
+
 # [v0.19.1](https://github.com/upb-uc4/University-Credits-4.0/compare/admission-v0.17.1...admission-v0.19.1) (2021-03-03)
 ## Feature
  - Added Exam Admissions