+++ product_code/certificate_service/CHANGELOG.md
@@ -1,3 +1,9 @@
+# [WIP - v0.19.1](https://github.com/upb-uc4/University-Credits-4.0/compare/certificate-v0.18.1...certificate-v0.19.1) (2021-XX-XX)
+## Feature
+ - Update to HLF-API v0.18.0
+## Refactor
+## Bugfix
+
 # [v0.18.1](https://github.com/upb-uc4/University-Credits-4.0/compare/certificate-v0.16.2...certificate-v0.18.1) (2021-02-18)
 ## Feature
 ## Refactor